Add deaths and kills counter, last may be broken -_-

+++ core/src/ee/taltech/iti0202/gui/game/networking/server/entity/PlayerEntity.java
@@ -32,8 +32,8 @@ public class PlayerEntity extends Entity implements Serializable {
     public MultiplayerAnimation animation;
     public boolean flippedAnimation = false;
 
-    public int kills = 0;
-    public int deaths = 0;
+    public short kills = 0;
+    public short deaths = 0;
 
     @Override
     public String toString() {
